+++ assets/src/scripts/monitoring-location/components/hydrograph/parameters.js
@@ -1,3 +1,6 @@
+// Required to initialize USWDS components after page load (WaterAlert ToolTips)
+import components from 'uswds/src/js/components';
+
 import {line} from 'd3-shape';
 import {select} from 'd3-selection';
 
@@ -173,11 +176,14 @@ export const plotSeriesSelectTable = function(elem,
             tr.append('td')
                 .append('a')
                     .attr('href', param => `${config.WATERALERT_SUBSCRIPTION}/?site_no=${siteno}&parm=${param.parameterCode}`)
+                    .attr('class', 'usa-tooltip usa-link')
                     .attr('data-position', 'left')
                     .attr('data-classes', 'width-full tablet:width-auto')
                     .attr('title', 'Subscribe to text or email alerts based on thresholds that you set')
                     .text('Subscribe');
         });
+    // Activate the USWDS toolTips for WaterAlert subscriptions
+    components.tooltip.init(elem.node());
 
     table.property('scrollTop', scrollTop);